lib.rs
// Import necessary dependencies
use ink_lang::contract;

// Define the EduChain contract
#[contract]
mod edu_chain {
    use ink_prelude::vec::Vec;
    use ink_storage::collections::HashMap;
    use ink_storage::traits::{PackedLayout, SpreadLayout};

    // Define the EduChain struct
    #[derive(Debug, PackedLayout, SpreadLayout)]
    struct EduChain {
        // Define state variables (e.g., HashMaps to store records and certifications)
        students: HashMap<AccountId, Vec<Certification>>,
        educators: HashMap<AccountId, Vec<Course>>,
    }

    // Define the Certification struct
    #[derive(Debug, PackedLayout, SpreadLayout)]
    struct Certification {
        // Define certification properties (e.g., name, date, issuer)
        name: String,
        date: u64,
        issuer: AccountId,
    }

    // Define the Course struct
    #[derive(Debug, PackedLayout, SpreadLayout)]
    struct Course {
        // Define course properties (e.g., name, description, educator)
        name: String,
        description: String,
        educator: AccountId,
    }

    // Implement functions for interacting with EduChain
    impl EduChain {
        // Constructor function
        #[ink(constructor)]
        fn new() -> Self {
            // Initialize EduChain state
            Self {
                students: HashMap::new(),
                educators: HashMap::new(),
            }
        }

        // Add a student's certification
        #[ink(message)]
        fn add_certification(&mut self, student: AccountId, name: String, date: u64, issuer: AccountId) {
            // Add certification to the student's record
            let certification = Certification { name, date, issuer };
            self.students.entry(student).or_insert(Vec::new()).push(certification);
        }

        // Add a course by an educator
        #[ink(message)]
        fn add_course(&mut self, educator: AccountId, name: String, description: String) {
            // Add the course to the educator's portfolio
            let course = Course { name, description, educator };
            self.educators.entry(educator).or_insert(Vec::new()).push(course);
        }

        // Get a student's certifications
        #[ink(message)]
        fn get_student_certifications(&self, student: AccountId) -> Vec<Certification> {
            // Retrieve and return the certifications of a student
            self.students.get(&student).cloned().unwrap_or_default()
        }

        // Get an educator's courses
        #[ink(message)]
        fn get_educator_courses(&self, educator: AccountId) -> Vec<Course> {
            // Retrieve and return the courses taught by an educator
            self.educators.get(&educator).cloned().unwrap_or_default()
        }
    }
}
